requirements.txt
argon2-cffi==18.3.0
bcrypt==3.1.4
cairocffi==0.9.0
CairoSVG==2.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
crispy-forms-foundation==0.7.0
cssselect2==0.2.1
defusedxml==0.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.3
django-appconf==1.0.2
django-bootstrap4==0.0.7
django-bootstrap4-datetimepicker==4.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-filter==2.0.0
django-heroku==0.3.1
django-mailgun==0.9.1
django-select2==6.3.1
django-select2-forms==2.0.2
django-sortedm2m==1.5.0
django-stronghold==0.3.0
django-tables2==2.0.2
django-widget-tweaks==1.4.3
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.11.3
idna==2.7
oauthlib==2.1.0
olefile==0.46
pdfkit==0.6.1
pdfrw==0.4
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycparser==2.19
PyJWT==1.6.4
PyPDF2==1.26.0
Pyphen==0.9.5
python-decouple==3.1
python-http-client==3.1.0
python-memcached==1.59
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.7
reportlab==3.5.9
requests==2.20.0
requests-oauthlib==1.0.0
sendgrid==5.6.0
sentry-sdk==0.6.4
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
sqlparse==0.2.4
static3==0.7.0
tinycss2==0.6.1
urllib3==1.24.1
webencodings==0.5.1
whitenoise==4.1.2
